Fix menu header bug in firefox

+++ config/unicorn.rb
@@ -1,42 +1,46 @@
-# Set your full path to application.
-app_dir = File.expand_path('../../', __FILE__)
-shared_dir = File.expand_path('../../../shared/', __FILE__)
+if Rails.env == 'production'
 
-# Set unicorn options
-worker_processes 2
-preload_app true
-timeout 30
+  # Set your full path to application.
+  app_dir = File.expand_path('../../', __FILE__)
+  shared_dir = File.expand_path('../../../shared/', __FILE__)
 
-# Fill path to your app
-working_directory app_dir
+  # Set unicorn options
+  worker_processes 2
+  preload_app true
+  timeout 30
 
-# Set up socket location
-listen "#{shared_dir}/sockets/unicorn.sock", :backlog => 64
+  # Fill path to your app
+  working_directory app_dir
 
-# Loging
-stderr_path "#{shared_dir}/log/unicorn.stderr.log"
-stdout_path "#{shared_dir}/log/unicorn.stdout.log"
+  # Set up socket location
+  listen "#{shared_dir}/sockets/unicorn.sock", :backlog => 64
 
-# Set master PID location
-pid "#{shared_dir}/pids/unicorn.pid"
+  # Loging
+  stderr_path "#{shared_dir}/log/unicorn.stderr.log"
+  stdout_path "#{shared_dir}/log/unicorn.stdout.log"
 
-before_fork do |server, worker|
-  defined?(ActiveRecord::Base) and ActiveRecord::Base.connection.disconnect!
-  old_pid = "#{server.config[:pid]}.oldbin"
-  if File.exists?(old_pid) && server.pid != old_pid
-    begin
-      sig = (worker.nr + 1) >= server.worker_processes ? :QUIT : :TTOU
-      Process.kill(sig, File.read(old_pid).to_i)
-    rescue Errno::ENOENT, Errno::ESRCH
-      # someone else did our job for us
+  # Set master PID location
+  pid "#{shared_dir}/pids/unicorn.pid"
+
+  before_fork do |server, worker|
+    defined?(ActiveRecord::Base) and ActiveRecord::Base.connection.disconnect!
+    old_pid = "#{server.config[:pid]}.oldbin"
+    if File.exists?(old_pid) && server.pid != old_pid
+      begin
+        sig = (worker.nr + 1) >= server.worker_processes ? :QUIT : :TTOU
+        Process.kill(sig, File.read(old_pid).to_i)
+      rescue Errno::ENOENT, Errno::ESRCH
+        # someone else did our job for us
+      end
     end
   end
-end
 
-after_fork do |server, worker|
-  defined?(ActiveRecord::Base) and ActiveRecord::Base.establish_connection
-end
+  after_fork do |server, worker|
+    defined?(ActiveRecord::Base) and ActiveRecord::Base.establish_connection
+  end
 
-before_exec do |server|
-  ENV['BUNDLE_GEMFILE'] = "#{app_dir}/Gemfile"
+  before_exec do |server|
+    ENV['BUNDLE_GEMFILE'] = "#{app_dir}/Gemfile"
+  end
+
 end
